pyproject.toml
[project]
name = "cu-project-2"
version = "0.1.0"
description = "Weather app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
